+++ io/user-input/UserInputCMD.h
@@ -29,7 +29,7 @@ public:
 
     bool addDirectedBridges();
 
-    int getNumberOfInterconnections();
+    long long getNumberOfInterconnections();
 
     Topology *getTopology();
